# [ENH] improve perf for `collapse_levels`; avoid explicit copy in `pivot_wider`