Please describe the changes proposed in the pull request:

- improve performance for `collapse_levels` especially if all levels in the MultiIndex are strings, and there are no empty strings
- avoid explicit copy in `pivot_wider`, since `pd.pivot` already makes a copy

**This PR relates to #1250 .**

Performance comparison this PR vs dev: 

dataframe with relatively small number of columns : [source](https://pandas.pydata.org/docs/user_guide/advanced.html#using-slicers)

```py

def mklbl(prefix, n):
    return ["%s%s" % (prefix, i) for i in range(n)]


miindex = pd.MultiIndex.from_product(
    [mklbl("A", 4), mklbl("B", 2), mklbl("C", 4), mklbl("D", 2)]
)


micolumns = pd.MultiIndex.from_tuples(
    [("a", "foo"), ("a", "bar"), ("b", "foo"), ("b", "bah")], names=["lvl0", "lvl1"]
)


dfmi = (
    pd.DataFrame(
        np.arange(len(miindex) * len(micolumns)).reshape(
            (len(miindex), len(micolumns))
        ),
        index=miindex,
        columns=micolumns,
    )
    .sort_index()
    .sort_index(axis=1)
)


m = dfmi.T

print(m.columns.size)
64

# this PR:
%timeit m.collapse_levels()
414 µs ± 8.14 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)

# dev:
%timeit m.collapse_levels()
84.7 µs ± 531 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```

- dataframe with a lot more columns:
```py

def mklbl(prefix, n):
    return ["%s%s" % (prefix, i) for i in range(n)]


miindex = pd.MultiIndex.from_product(
    [mklbl("A", 40), mklbl("B", 20), mklbl("C", 40), mklbl("D", 20)]
)


micolumns = pd.MultiIndex.from_tuples(
    [("a", "foo"), ("a", "bar"), ("b", "foo"), ("b", "bah")], names=["lvl0", "lvl1"]
)


dfmi = (
    pd.DataFrame(
        np.arange(len(miindex) * len(micolumns)).reshape(
            (len(miindex), len(micolumns))
        ),
        index=miindex,
        columns=micolumns,
    )
    .sort_index()
    .sort_index(axis=1)
)


m = dfmi.T

print(m.columns.size)
64000

# this PR:
%timeit m.collapse_levels()
27.2 ms ± 250 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

# dev:
%timeit m.collapse_levels()
48.7 ms ± 211 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

Test when columns have integers: 

```py
m.columns = [m.columns.get_level_values(n) for n in range(3)] + [np.arange(1, m.columns.size + 1)]
38.5 ms ± 487 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

# dev:
%timeit m.collapse_levels()
55.3 ms ± 2.82 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

add a float column:

```py
m = dfmi.T
m.columns = [m.columns.get_level_values(n) for n in range(2)] + [np.arange(1, m.columns.size + 1), np.arange(1, m.columns.size + 1).astype(float)]

# PR:
%timeit m.collapse_levels()
44.9 ms ± 203 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

# dev
%timeit m.collapse_levels()
66.6 ms ± 461 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

- @ericmjl